Code cleaup: Remove unnecessary constant.
Because useNQT() isn't necessary, there's no need to keep its constant either.

Line 296 doesn't need this because NQT_BLOCK is zero, so it'll only ever trigger this if the chain height is zero, but that should not be an issue.

Line 338 will never get triggered because it will never be lower than zero.